42_cursus/Transcendence/projet_transcendence/rendu_transcendence/lets_see/requirements.txt
asgiref==3.8.1
channels==4.2.0
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-allauth==65.3.1
django-channels==0.7.0
daphne==4.1.2
PyJWT==2.10.1
pillow==11.1.0
psycopg2-binary==2.9.10
requests==2.32.3
channels-redis==4.2.1
sqlparse==0.5.2
#gunicorn==23.0.0
#djangorestframework_simplejwt==5.4.0

django-two-factor-auth==1.17.0
django-otp==1.6.0
phonenumbers==9.0.3
qrcode==7.4.2